Update proposal on this Q&A
I mentioned before Revolutionary tool/method guide might be interesting to link (as a simple 'push to do it' method for S-OFF and +) in addition to the current link (that is a good description for the issue, besides Unrevokable tool/method) but I realized all this is in constant evolution with new phone-models/methods. Now I'm proposing just naming additional most notorious methods (simple and not) nowadays, for interested users to search about.
I'm using this file to propose the modification but I'm not sure this is the right, or unique file to do it upon, just notice.