Fix binary to library trick
The binaries are now compiled using gradle "externalNativeBuild".
The binaries are built into `AdAway/buird/intermediates/ndbBuild` and
copied with a SO name into `AdAway/libs` in order to be packed into the
apk.
This also fix the variant of included binaries: debug binaries are now
shipped in the debug apk.