Allow installing debug builds
Setting the debug signing config to null causes us to output
unsigned APKs when building on the command line, which in turn
breaks the installDebug task, making it unavailable. Setting
release signing config to null serves no purpose either because
the output is going to always be an unsigned APK. Skipping
the assignment defaults debug signing config to the system debug
keystore which unbreaks the installDebug task, allowing for easier
development.

Signed-off-by: Harsh Shandilya <msfjarvis@gmail.com>