Update AdwareLiveData.java : Fix potential NullPointerException in createInstallFromPackageInfo
Prevent a possible NullPointerException in createInstallFromPackageInfo by adding a null check for packageInfo.applicationInfo. If applicationInfo is null, we now return a default application name ("Unknown") instead of calling getApplicationLabel, which would otherwise throw an exception. This improves the stability of the app and prevents crashes due to unexpected null values in PackageInfo.